requirements.txt
six
mongoengine
redis
django>=1.9
cerberus
requests
celery==3.1.23
django-celery==3.1.17
django-kombu
MySQL-python
ConcurrentLogHandler
pymongo
python-dateutil
future
fabric